examples of problem solving when given a prompt
These are from codewars. A prompt is given and i have to use python code to solve it. The code has been tested in both an IDE (PyCharm) and in their developer environment. This is all code i wrote. I only compared my code to the solution after it was submitted (or i was very stumped, as the instructions can lack details).